requirements.txt
uvicorn[standard]==0.34.2
fastapi==0.115.12
celery==5.5.2
redis==6.0.0
spacy==3.8.5
ru_core_news_md @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.8.0/ru_core_news_md-3.8.0-py3-none-any.whl#sha256=43f456d40e4b70726874e01ce01fce32f50089ba98a2660292e9a6c63a06ebe9
aiofiles==24.1.0
keyphrase-vectorizers==0.0.13
frontend==0.0.3
pymorphy3~=2.0.3
numpy~=2.0.1
pydantic~=2.11.4
pydantic-settings==2.9.1
aiohttp~=3.11.18
pytesseract~=0.3.13
pillow~=11.1.0
scikit-learn~=1.6.1
RapidFuzz~=3.13.0
sentence-transformers~=4.1.0
fitz~=0.0.1.dev2
PyMuPDF==1.25.5
loguru~=0.7.3